+++ src/input_streams.rs
@@ -330,6 +330,38 @@ impl<'a> InputStreams<'a> {
             UserInput::VirtualDPad { .. } => {
                 self.input_axis_pair(input).unwrap_or_default().length()
             }
+            UserInput::Chord(inputs) => {
+                let mut value = 0.0;
+                let mut has_axis = false;
+
+                // Prioritize axis over button input values
+                for input in inputs.iter() {
+                    value += match input {
+                        InputKind::SingleAxis(axis) => {
+                            has_axis = true;
+                            self.input_value(&UserInput::Single(InputKind::SingleAxis(*axis)), true)
+                        }
+                        InputKind::MouseWheel(axis) => {
+                            has_axis = true;
+                            self.input_value(&UserInput::Single(InputKind::MouseWheel(*axis)), true)
+                        }
+                        InputKind::MouseMotion(axis) => {
+                            has_axis = true;
+                            self.input_value(
+                                &UserInput::Single(InputKind::MouseMotion(*axis)),
+                                true,
+                            )
+                        }
+                        _ => 0.0,
+                    }
+                }
+
+                if has_axis {
+                    return value;
+                }
+
+                use_button_value()
+            }
             // This is required because upstream bevy::input still waffles about whether triggers are buttons or axes
             UserInput::Single(InputKind::GamepadButton(button_type)) => {
                 if let Some(gamepad) = self.guess_gamepad() {